Fix bug #2241381 (segmentation fault when upgrading tavern to inn). The savegame now fails to load with an exception describing what is wrong with it, instead of using the out-of-range ware index found in the file and crashing. (Just wondering how the savegame could become broken...) Add some asserts to detect out-of-range ware/worker index. Add some more input validation to Request::Read, such as count being positive. In Request, change the type of m_count, m_required_time, m_required_interval, m_last_request_time from int32_t to uint32_t because they were read and written as unsigned types and doing silent conversions to a signed type (for seemingly no reason) will probably cause bugs.
git-svn-id: https://widelands.svn.sourceforge.net/svnroot/widelands/trunk@3496 37b2a8de-5219-0410-9f54-a31bc463ab9c

+++ src/request.h
@@ -73,12 +73,12 @@ struct Request : public Trackable {
 	Ware_Index get_index() const {return m_index;}
 	int32_t get_type() const {return m_type;}
 	bool is_idle() const {return m_idle;}
-	int32_t get_count() const {return m_count;}
+	uint32_t get_count() const {return m_count;}
 	bool is_open() const
-	{return m_idle || m_count > static_cast<int32_t>(m_transfers.size());}
+	{return m_idle || m_count > m_transfers.size();}
 	Economy * get_economy() const throw () {return m_economy;}
-	int32_t get_required_time();
-	int32_t get_last_request_time() {return m_last_request_time;}
+	uint32_t get_required_time();
+	uint32_t get_last_request_time() {return m_last_request_time;}
 	int32_t get_priority(int32_t cost);
 
 	Flag * get_target_flag();
@@ -120,14 +120,14 @@ private:
 	Economy         * m_economy;
 	Ware_Index        m_index;             //  the index of the ware descr
 	bool              m_idle;
-	int32_t               m_count;             //  how many do we need in total
+	uint32_t          m_count;             //  how many do we need in total
 
 	callback_t        m_callbackfn;        //  called on request success
 	void            * m_callbackdata;
 
-	int32_t m_required_time; //  when do we need the first ware (can be in the past)
-	int32_t               m_required_interval; //  time between items
-	int32_t m_last_request_time;
+	uint32_t          m_required_time; //  when do we need the first ware (can be in the past)
+	uint32_t          m_required_interval; //  time between items
+	uint32_t          m_last_request_time;
 
 	TransferList      m_transfers;         //  maximum size is m_count
